feat: the testing of the line reversal algorithm has been implemented
